# docs/apispec.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "MyoHanMeeting Backend",
        "description": "유기묘 입양 서비스 묘한만남의 백엔드 서버 문서입니다.",
        "contact": {
            "email": "bomlee427@gmail.com"
        },
        "version": "v1"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
...
    "paths": {
...
        "/v1/adoption/applications": {
            "post": {
                "tags": [
                    "Adopt Application"
                ],
                "summary": "분양신청 작성",
                "description": "분양 신청서를 작성합니다.",
                "operationId": "createApplication",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/adoptApplicationForm"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "400": {
                        "description": "잘못된 요청입니다.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "status": "400 BAD_REQUEST",
                                    "timestamp": "2023-06-10T09:19:08.550Z",
                                    "message": "INVALID PARAMETER",
                                    "debugMessage": ""
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "서버측 오류입니다.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "status": "500 INTERNAL_SERVER_ERROR",
                                    "timestamp": "2023-06-10T09:19:08.550Z",
                                    "message": "SERVER ERROR",
                                    "debugMessage": ""
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "작성 성공",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommonResponseDto"
                                },
                                "example": {
                                    "status": "200 OK",
                                    "timestamp": "2023-06-10T09:19:08.550Z",
                                    "message": "SUCCESS",
                                    "data": {
                                        "applicationId": 1
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "인증이 필요한 URI입니다.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "status": "401 UNAUTHORIZED",
                                    "timestamp": "2023-06-10T09:19:08.550Z",
                                    "message": "AUTHORIZATION REQUIRED",
                                    "debugMessage": ""
                                }
                            }
                        }
                    },
                    "405": {
                        "description": "잘못된 메서드 또는 URI 요청입니다.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "status": "405 METHOD_NOT_ALLOWED",
                                    "timestamp": "2023-06-10T09:19:08.550Z",
                                    "message": "METHOD_NOT_ALLOWED",
                                    "debugMessage": ""
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "요청한 자원을 찾을 수 없습니다.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "status": "404 NOT_FOUND",
                                    "timestamp": "2023-06-10T09:19:08.550Z",
                                    "message": "NO RESOURCE FOUND",
                                    "debugMessage": ""
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        },
...
        "/v1/adoption/applications/my": {
            "get": {
                "tags": [
                    "Adopt Application"
                ],
                "summary": "내가 한 분양신청 목록조회",
                "description": "자신이 작성한 분양신청 목록을 조회합니다.",
                "operationId": "getMyApplicationList",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "페이지를 설정합니다.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "리미트를 설정합니다.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 100
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "정렬 기준을 선택합니다.<br><br>, 를 구분자로 항목 앞에 -를 붙이면 내림차순, 붙이지 않으면 오름차순 정렬입니다.<br>정렬 가능한 항목은 다음과 같습니다: createdAt(작성일), applicationCount(신청수), commentCount(댓글수)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "-createdAt,applicationCount"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "잘못된 요청입니다.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "status": "400 BAD_REQUEST",
                                    "timestamp": "2023-06-10T09:19:08.550Z",
                                    "message": "INVALID PARAMETER",
                                    "debugMessage": ""
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "서버측 오류입니다.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "status": "500 INTERNAL_SERVER_ERROR",
                                    "timestamp": "2023-06-10T09:19:08.550Z",
                                    "message": "SERVER ERROR",
                                    "debugMessage": ""
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommonResponseDtoListAdoptApplication"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "인증이 필요한 URI입니다.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "status": "401 UNAUTHORIZED",
                                    "timestamp": "2023-06-10T09:19:08.550Z",
                                    "message": "AUTHORIZATION REQUIRED",
                                    "debugMessage": ""
                                }
                            }
                        }
                    },
                    "405": {
                        "description": "잘못된 메서드 또는 URI 요청입니다.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "status": "405 METHOD_NOT_ALLOWED",
                                    "timestamp": "2023-06-10T09:19:08.550Z",
                                    "message": "METHOD_NOT_ALLOWED",
                                    "debugMessage": ""
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        },
        "/v1/adoption/applications/byNotice/{noticeId}": {
            "get": {
                "tags": [
                    "Adopt Application"
                ],
                "summary": "특정 공고에 달린 신청 모아보기",
                "description": "개별 공고에 달린 신청을 한번에 모아봅니다. 해당 공고의 작성자만 열람 가능합니다.",
                "operationId": "getApplicationListByNotice",
                "parameters": [
                    {
                        "name": "noticeId",
                        "in": "path",
                        "description": "신청 목록을 조회하고자 하는 분양공고의 id입니다.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "페이지를 설정합니다.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "리미트를 설정합니다.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 100
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "정렬 기준을 선택합니다.<br><br>, 를 구분자로 항목 앞에 -를 붙이면 내림차순, 붙이지 않으면 오름차순 정렬입니다.<br>정렬 가능한 항목은 다음과 같습니다: createdAt(작성일), applicationCount(신청수), commentCount(댓글수)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "-createdAt,applicationCount"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "잘못된 요청입니다.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "status": "400 BAD_REQUEST",
                                    "timestamp": "2023-06-10T09:19:08.550Z",
                                    "message": "INVALID PARAMETER",
                                    "debugMessage": ""
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "서버측 오류입니다.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "status": "500 INTERNAL_SERVER_ERROR",
                                    "timestamp": "2023-06-10T09:19:08.550Z",
                                    "message": "SERVER ERROR",
                                    "debugMessage": ""
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommonResponseDtoListAdoptApplication"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "허가된 접근이 아닙니다.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "status": "403 BAD_REQUEST",
                                    "timestamp": "2023-06-10T09:19:08.550Z",
                                    "message": "ACCESS DENIED",
                                    "debugMessage": ""
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "인증이 필요한 URI입니다.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "status": "401 UNAUTHORIZED",
                                    "timestamp": "2023-06-10T09:19:08.550Z",
                                    "message": "AUTHORIZATION REQUIRED",
                                    "debugMessage": ""
                                }
                            }
                        }
                    },
                    "405": {
                        "description": "잘못된 메서드 또는 URI 요청입니다.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "status": "405 METHOD_NOT_ALLOWED",
                                    "timestamp": "2023-06-10T09:19:08.550Z",
                                    "message": "METHOD_NOT_ALLOWED",
                                    "debugMessage": ""
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "요청한 자원을 찾을 수 없습니다.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "status": "404 NOT_FOUND",
                                    "timestamp": "2023-06-10T09:19:08.550Z",
                                    "message": "NO RESOURCE FOUND",
                                    "debugMessage": ""
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorResponse": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string",
                        "description": "HTTP 상태 코드입니다.",
                        "enum": [
                            "200 OK",
                            "400 BAD_REQUEST",
                            "401 UNAUTHORIZED",
                            "403 FORBIDDEN",
                            "404 NOT_FOUND",
                            "405 METHOD_NOT_ALLOWED",
                            "500 INTERNAL_SERVER_ERROR"
                        ]
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "message": {
                        "type": "string",
                        "description": "응답에 대한 요약을 나타냅니다."
                    },
                    "debugMessage": {
                        "type": "string",
                        "description": "오류에 관한 상세 디버깅 메시지를 나타냅니다."
                    }
                }
            },
            "CommonResponseDto": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "200 OK"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "message": {
                        "type": "string",
                        "example": "SUCCESS"
                    },
                    "data": {
                        "type": "object"
                    }
                }
            },
...
            "Applicant": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "age": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "gender": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string"
                    },
                    "phoneNumber": {
                        "type": "string"
                    },
                    "job": {
                        "type": "string"
                    },
                    "married": {
                        "type": "string"
                    }
                }
            },
            "Survey": {
                "type": "object",
                "properties": {
                    "answer1_1": {
                        "type": "string"
                    },
                    "answer1_2": {
                        "type": "string"
                    },
                    "answer2_1": {
                        "type": "string"
                    },
                    "answer2_2": {
                        "type": "string"
                    },
                    "answer3": {
                        "type": "string"
                    },
                    "answer4": {
                        "type": "string"
                    },
                    "answer5": {
                        "type": "string"
                    },
                    "answer6": {
                        "type": "string"
                    }
                }
            },
            "adoptApplicationForm": {
                "type": "object",
                "properties": {
                    "noticeId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "applicant": {
                        "$ref": "#/components/schemas/Applicant"
                    },
                    "survey": {
                        "$ref": "#/components/schemas/Survey"
                    },
                    "content": {
                        "type": "string"
                    }
                }
            },
...
            "author": {
                "type": "object",
                "properties": {
                    "authorId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "nickname": {
                        "type": "string"
                    },
                    "profileImageUrl": {
                        "type": "string"
                    }
                }
            },
...
            "AdoptApplication": {
                "type": "object",
                "properties": {
                    "applicationId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "noticeId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "author": {
                        "$ref": "#/components/schemas/author"
                    },
                    "applicant": {
                        "$ref": "#/components/schemas/applicant"
                    },
                    "survey": {
                        "$ref": "#/components/schemas/survey"
                    },
                    "content": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "applicant": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "age": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "gender": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string"
                    },
                    "phoneNumber": {
                        "type": "string"
                    },
                    "job": {
                        "type": "string"
                    },
                    "married": {
                        "type": "string"
                    }
                }
            },
...
            "CommonResponseDtoListAdoptApplication": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "200 OK"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "message": {
                        "type": "string",
                        "example": "SUCCESS"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AdoptApplication"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "Authorization": {
                "type": "http",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}